separated the cytoscape graph elements generation and tooltip creation logic from the main file
